ai-engine/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq>=0.2.0
langchain-community
langgraph
tavily-python
pydantic
fastembed  